+++ src/command-title.ts
@@ -1,0 +1,8 @@
+export enum CommandTitle {
+  OpenInXCode = 'Open in Xcode',
+  OpenInAndroidStudio = 'Open in Android Studio',
+  RunForIOS = 'iOS',
+  RunForAndroid = 'Android',
+  Sync = 'Sync',
+  RunForWeb = 'Web',
+}